+++ slim/__init__.py
@@ -5,4 +5,4 @@ from . import base
 from . import support
 from . import utils
 
-__version__ = '0.4.11'
+__version__ = '0.4.12'